move v2 http endpoints to http.go file
